Fix: fix typescript strictNull error in contexts files

+++ context/BuskingContext.tsx
@@ -17,7 +17,7 @@ import { useAuthContext } from './AuthContext';
 import { useUserDataContext } from './UserDataContext';
 
 type ContextProps = {
-  buskingData: BuskingData | undefined;
+  buskingData: BuskingData | null;
   makeBusking: (buskingInform: BuskingInform) => Promise<void>;
   removeBuskingSong: (sid: string) => Promise<void>;
   removeBusking: () => Promise<void>;
@@ -37,14 +37,14 @@ type ContextProps = {
   ) => Promise<void>;
   syncBuskingData: (
     userId: string,
-    onUpdate: (value: BuskingData | undefined) => void
+    onUpdate: (value: BuskingData | null) => void
   ) => Unsubscribe;
-  getBuskingData: (userId: string) => Promise<BuskingData>;
+  getBuskingData: (userId: string) => Promise<BuskingData | null>;
   isbuskingDataLoading: boolean;
   applyBuskingSongAgain: (nowSong: ApplianceData) => Promise<void>;
 };
 
-const BuskingContext = createContext<ContextProps>(undefined);
+const BuskingContext = createContext<ContextProps>({} as ContextProps);
 
 type Props = {
   buskingRepository: BuskingRepository;
@@ -52,12 +52,14 @@ type Props = {
 };
 
 export function BuskingContextProvider({ buskingRepository, children }: Props) {
-  const [buskingData, setBuskingData] = useState<BuskingData | undefined>();
+  // useState 초기값 안넣으면 undefined 되는거 생각하기
+  const [buskingData, setBuskingData] = useState<BuskingData | null>(null);
   const { uid } = useAuthContext();
   const { userData } = useUserDataContext();
   const [isbuskingDataLoading, setIsbuskingDataLoading] =
     useState<boolean>(true);
   useEffect(() => {
+    if (!uid) throw new Error('no uid!!');
     setIsbuskingDataLoading(true);
     return buskingRepository.syncBuskingData(uid, (data) => {
       setBuskingData(data);
@@ -65,15 +67,19 @@ export function BuskingContextProvider({ buskingRepository, children }: Props) {
     });
   }, [uid, userData]);
 
+  //TODO: uid 어디서 받아와야하는지 고민해보기
   const makeBusking = async (buskingInform: BuskingInform) => {
+    if (!uid) throw new Error('no uid!!');
     return buskingRepository.makeBusking(uid, buskingInform);
   };
 
   const removeBuskingSong = async (sid: string) => {
+    if (!uid) throw new Error('no uid!!');
     return buskingRepository.removeBuskingSong(uid, sid);
   };
 
   const removeBusking = async () => {
+    if (!uid) throw new Error('no uid!!');
     return buskingRepository.removeBusking(uid);
   };
   const applyOldBuskingSong = async (
@@ -93,6 +99,7 @@ export function BuskingContextProvider({ buskingRepository, children }: Props) {
   };
 
   const applyBuskingSongAgain = async (nowSong: ApplianceData) => {
+    if (!uid) throw new Error('no uid!!');
     return buskingRepository.applyBuskingSongAgain(uid, nowSong, nowSong.sid);
   };
 
@@ -114,7 +121,7 @@ export function BuskingContextProvider({ buskingRepository, children }: Props) {
 
   const syncBuskingData = (
     userId: string,
-    onUpdate: (value: BuskingData | undefined) => void
+    onUpdate: (value: BuskingData | null) => void
   ) => {
     return buskingRepository.syncBuskingData(userId, onUpdate);
   };
